requirements.txt
timm==0.6.7
torch==2.0.1
plyfile==0.8.1
diffusers==0.23.1
peft==0.6.2
diffusers==0.30.0
einops==0.8.0
accelerate==0.28.0
tensorflow==2.11.0
waymo-open-dataset-tf-2-11-0==1.6.1
torchvision 
scipy
numpy
loguru
tensorboard
imageio[ffmpeg]
opencv-python
Pillow
open3d
gradio
gsplat